1. 修改代码命名空间 2. 添加对spring boot 的支持
