+++ rate-limiter-client/src/main/java/cn/caijiajia/ratelimiter/client/RateLimiterConstants.java
@@ -1,0 +1,13 @@
+package cn.caijiajia.ratelimiter.client;
+
+/**
+ * @author wukaiqiang
+ */
+public class RateLimiterConstants {
+    public static final String RATE_LIMITER_KEY_PREFIX = "rate_limiter:";
+
+
+    public static final String RATE_LIMITER_INIT_METHOD = "init";
+    public static final String RATE_LIMITER_DELETE_METHOD = "delete";
+    public static final String RATE_LIMITER_ACQUIRE_METHOD = "acquire";
+}
