--- rate-limiter-client/src/main/java/cn/wukq/ratelimiter/client/RateLimiterConstants.java
@@ -1,13 +1,0 @@
-package cn.wukq.ratelimiter.client;
-
-/**
- * @author wukaiqiang
- */
-public class RateLimiterConstants {
-    public static final String RATE_LIMITER_KEY_PREFIX = "rate_limiter:";
-
-
-    public static final String RATE_LIMITER_INIT_METHOD = "init";
-    public static final String RATE_LIMITER_DELETE_METHOD = "delete";
-    public static final String RATE_LIMITER_ACQUIRE_METHOD = "acquire";
-}
