+++ rate-limiter-client/src/main/resources/applicationContext-rateLimiter.xml
@@ -1,15 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.springframework.org/schema/beans"
+       xmlns:p="http://www.springframework.org/schema/p"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
     <bean id="rateLimiterLua" class="org.springframework.data.redis.core.script.DefaultRedisScript">
         <property name="location" value="classpath:rate_limiter.lua"/>
         <property name="resultType" value="java.lang.Long"/>
     </bean>
 
-    <bean id="rateLimiterRedisClient" class="cn.wukq.ratelimiter.client.RedisClient">
-        <property name="hostName" value="${redis.host.user.rateLimiter}"/>
-        <property name="port" value="${redis.port.user.rateLimiter}"/>
-        <property name="database" value="${redis.database.user.rateLimiter}"/>
+    <bean id="poolConfig" class="redis.clients.jedis.JedisPoolConfig">
+        <!--可以再次自定义redis连接池参数-->
+    </bean>
+
+    <bean id="connectionFactory"
+          class="org.springframework.data.redis.connection.jedis.JedisConnectionFactory"
+          p:host-name="${rate.limiter.redis.host}"
+          p:port="${rate.limiter.redis.port}"
+          p:password="${rate.limiter.redis.pass}"
+          p:pool-config-ref="poolConfig"/>
+
+    <bean id="stringRedisTemplate" class="org.springframework.data.redis.core.StringRedisTemplate">
+        <property name="connectionFactory" ref="connectionFactory"/>
+        <property name="defaultSerializer">
+            <bean class="org.springframework.data.redis.serializer.StringRedisSerializer"/>
+        </property>
+    </bean>
+
+    <bean id="rateLimiterClient" class="cn.caijiajia.ratelimiter.client.RateLimiterClient">
+        <constructor-arg index="0" ref="stringRedisTemplate"/>
+        <constructor-arg index="1" ref="rateLimiterLua"/>
     </bean>
 
 </beans>
