+++ rate-limiter-server/src/main/java/cn/caijiajia/ratelimiter/server/vo/RateLimiterVo.java
@@ -1,0 +1,23 @@
+package cn.caijiajia.ratelimiter.server.vo;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * @author wukaiqiang
+ */
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RateLimiterVo {
+    private String name;
+    private String apps;
+    private Integer maxPermits;
+    private Integer rate;
+    private Integer currPermits;
+    private String lastPermitTimestamp;
+
+}
