+++ rate-limiter-starter/pom.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>rate-limiter</artifactId>
+        <groupId>cn.caijiajia</groupId>
+        <version>1.0.0.SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>rate-limiter-starter</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>cn.caijiajia</groupId>
+            <artifactId>rate-limiter-autoConfigure</artifactId>
+        </dependency>
+    </dependencies>
+
+</project>
